Update expects in antialiased tests
Previously, when `outputDiffMask` was `false`, `diffOutput` was assign
the same value as `img1`. This wasn't deep copy, so both variables were
pointing to the same image data. This was usually OK, because we don't
look back at pixels that were already compared. The exception is the
case of `antialiasing = true`. In this case we do look back at
previously compared pixels and since they were modified, this changed
the result of comparison.

This was fixed in previous commit so that `diffOutput` is always a copy
of input image. This is why the expects were changed as well.